refactor: :truck: moved `viewContactModal` and `StatusTag` to different files and added `id` field in table
