+++ src/pages/contacts/index.tsx
@@ -1,12 +1,13 @@
-import { useState, useEffect, FC } from "react";
-import { Table, Tag, TablePaginationConfig, Modal, Typography, Button } from "antd";
-import { PresetColorType } from "antd/lib/_util/colors";
+import { useState, useEffect } from "react";
+import { Table, TablePaginationConfig } from "antd";
 import { useAppDispatch } from "../../hooks/useAddDispatch";
 import { useAppSelector } from "../../hooks/useAppSelector";
 import { EContactStatus } from "../../types/enum";
 import { TRootState } from "../../types/types";
 import { fetchAllContacts } from "../../actions/contacts.action";
 import { IContact } from "../../types/interface";
+import StatusTag from "./statusTag";
+import ViewContactModal from "./viewContactModal";
 
 function Contacts() {
 	const dispatch = useAppDispatch();
@@ -51,37 +52,12 @@ function Contacts() {
 		};
 	};
 
-	const closeModalHandler = () => {
-		setModalState((prevState) => ({ ...prevState, visible: false }));
-	};
-
-	interface StatusProps {
-		/**
-		 * Status of the contact
-		 */
-		status: EContactStatus;
-
-		/**
-		 * Additional class name
-		 */
-		className?: string;
-	}
-
-	const Status: FC<StatusProps> = ({ status, className }) => {
-		let color: PresetColorType = "red";
-
-		if (status === EContactStatus["Pending"]) color = "red";
-		else if (status === EContactStatus["In Progress"]) color = "gold";
-		else if (status === EContactStatus["Completed"]) color = "green";
-
-		return (
-			<Tag color={color} className={className}>
-				{status}
-			</Tag>
-		);
-	};
-
 	const columns = [
+		{
+			title: "ID",
+			dataIndex: "id",
+			key: "id",
+		},
 		{
 			title: "Name",
 			dataIndex: "name",
@@ -102,7 +78,7 @@ function Contacts() {
 			dataIndex: "status",
 			key: "status",
 			render: (status: EContactStatus) => {
-				return <Status status={status} />;
+				return <StatusTag status={status} />;
 			},
 		},
 		{
@@ -114,60 +90,7 @@ function Contacts() {
 
 	return (
 		<main className="px-16 py-10">
-			<Modal
-				title={
-					<div className="flex items-center">
-						<Typography.Title level={5} className="!m-0 !mb-0 !mt-0">
-							{modalState.record.name}
-						</Typography.Title>
-						<Status status={modalState.record.status} className="!ml-3" />
-					</div>
-				}
-				open={modalState.visible}
-				onCancel={closeModalHandler}
-				closable
-			>
-				<div className="flex items-center mt-3">
-					<Typography.Title level={4} className="!mb-0">
-						Email:-
-					</Typography.Title>
-					<Typography.Title level={5} className="ml-4 !mt-0 !mb-0">
-						{modalState.record.email}
-					</Typography.Title>
-				</div>
-
-				<div className="flex items-center mt-3">
-					<Typography.Title level={4} className="!mb-0">
-						Phone:-
-					</Typography.Title>
-					<Typography.Title level={5} className="ml-4 !mt-0 !mb-0">
-						{modalState.record.phone}
-					</Typography.Title>
-				</div>
-
-				<div className="flex items-center mt-3">
-					<Typography.Title level={4} className="!mb-0">
-						Subject:-
-					</Typography.Title>
-					<Typography.Title level={5} className="ml-4 !mt-0 !mb-0">
-						{modalState.record.subject}
-					</Typography.Title>
-				</div>
-
-				<div className="flex items-center mt-3">
-					<Typography.Title level={4} className="!mb-0">
-						Message:-
-					</Typography.Title>
-					<Typography.Paragraph className="ml-4 !mt-0 !mb-0">
-						{modalState.record.message}
-					</Typography.Paragraph>
-				</div>
-
-				<div className="mt-12">
-					<Button>Update Status</Button>
-					<Button className="ml-3">Delete</Button>
-				</div>
-			</Modal>
+			<ViewContactModal modalState={modalState} setModalState={setModalState} />
 
 			<Table
 				loading={getLoading}
